fix: adopt macros and as with the queue contract, several structures were moved to msg.rs
