+++ contracts/query-queue/examples/schema.rs
@@ -1,25 +1,11 @@
-use std::env::current_dir;
-use std::fs::create_dir_all;
+use cosmwasm_schema::write_api;
 
-use cosmwasm_schema::{export_schema, remove_schemas, schema_for};
-
-use query_queue::contract::{
-    CountResponse, ListResponse, RawResponse, ReducerResponse, SumResponse,
-};
 use query_queue::msg::{ExecuteMsg, InstantiateMsg, QueryMsg};
 
 fn main() {
-    let mut out_dir = current_dir().unwrap();
-    out_dir.push("schema");
-    create_dir_all(&out_dir).unwrap();
-    remove_schemas(&out_dir).unwrap();
-
-    export_schema(&schema_for!(InstantiateMsg), &out_dir);
-    export_schema(&schema_for!(ExecuteMsg), &out_dir);
-    export_schema(&schema_for!(QueryMsg), &out_dir);
-    export_schema(&schema_for!(RawResponse), &out_dir);
-    export_schema(&schema_for!(CountResponse), &out_dir);
-    export_schema(&schema_for!(SumResponse), &out_dir);
-    export_schema(&schema_for!(ReducerResponse), &out_dir);
-    export_schema(&schema_for!(ListResponse), &out_dir);
+    write_api! {
+        instantiate: InstantiateMsg,
+        query: QueryMsg,
+        execute: ExecuteMsg,
+    }
 }
