--- contracts/query-queue/schema/count_response.json
@@ -1,15 +1,0 @@
-{
-  "$schema": "http://json-schema.org/draft-07/schema#",
-  "title": "CountResponse",
-  "type": "object",
-  "required": [
-    "count"
-  ],
-  "properties": {
-    "count": {
-      "type": "integer",
-      "format": "uint32",
-      "minimum": 0.0
-    }
-  }
-}
