--- contracts/query-queue/schema/execute_msg.json
@@ -1,26 +1,0 @@
-{
-  "$schema": "http://json-schema.org/draft-07/schema#",
-  "title": "ExecuteMsg",
-  "oneOf": [
-    {
-      "type": "object",
-      "required": [
-        "change_address"
-      ],
-      "properties": {
-        "change_address": {
-          "type": "object",
-          "required": [
-            "queue_address"
-          ],
-          "properties": {
-            "queue_address": {
-              "type": "string"
-            }
-          }
-        }
-      },
-      "additionalProperties": false
-    }
-  ]
-}
