--- contracts/query-queue/schema/instantiate_msg.json
@@ -1,13 +1,0 @@
-{
-  "$schema": "http://json-schema.org/draft-07/schema#",
-  "title": "InstantiateMsg",
-  "type": "object",
-  "required": [
-    "queue_address"
-  ],
-  "properties": {
-    "queue_address": {
-      "type": "string"
-    }
-  }
-}
