--- contracts/query-queue/schema/raw_response.json
@@ -1,14 +1,0 @@
-{
-  "$schema": "http://json-schema.org/draft-07/schema#",
-  "title": "RawResponse",
-  "type": "object",
-  "properties": {
-    "item": {
-      "type": [
-        "integer",
-        "null"
-      ],
-      "format": "int32"
-    }
-  }
-}
