--- contracts/query-queue/schema/reducer_response.json
@@ -1,28 +1,0 @@
-{
-  "$schema": "http://json-schema.org/draft-07/schema#",
-  "title": "ReducerResponse",
-  "type": "object",
-  "required": [
-    "counters"
-  ],
-  "properties": {
-    "counters": {
-      "type": "array",
-      "items": {
-        "type": "array",
-        "items": [
-          {
-            "type": "integer",
-            "format": "int32"
-          },
-          {
-            "type": "integer",
-            "format": "int32"
-          }
-        ],
-        "maxItems": 2,
-        "minItems": 2
-      }
-    }
-  }
-}
