--- contracts/query-queue/schema/sum_response.json
@@ -1,14 +1,0 @@
-{
-  "$schema": "http://json-schema.org/draft-07/schema#",
-  "title": "SumResponse",
-  "type": "object",
-  "required": [
-    "sum"
-  ],
-  "properties": {
-    "sum": {
-      "type": "integer",
-      "format": "int32"
-    }
-  }
-}
